+++ packages/platform-apple/src/runner/__tests__/runner-cache-metadata.test.ts
@@ -1,0 +1,150 @@
+import { test } from 'vitest';
+import assert from 'node:assert/strict';
+import { IOS_DEVICE, IOS_SIMULATOR, MACOS_DEVICE } from './device-fixtures.ts';
+import {
+  resolveRunnerBundleBuildSettings,
+  resolveRunnerMaxConcurrentDestinationsFlag,
+  resolveRunnerSigningBuildSettings,
+  resolveRunnerPerformanceBuildSettings,
+  resolveRunnerSandboxBuildArgs,
+} from '../runner-cache-metadata.ts';
+
+test('resolveRunnerMaxConcurrentDestinationsFlag uses simulator flag for simulators', () => {
+  assert.equal(
+    resolveRunnerMaxConcurrentDestinationsFlag(IOS_SIMULATOR),
+    '-maximum-concurrent-test-simulator-destinations',
+  );
+});
+
+test('resolveRunnerMaxConcurrentDestinationsFlag uses device flag for physical devices', () => {
+  assert.equal(
+    resolveRunnerMaxConcurrentDestinationsFlag(IOS_DEVICE),
+    '-maximum-concurrent-test-device-destinations',
+  );
+});
+
+test('resolveRunnerMaxConcurrentDestinationsFlag uses device flag for macOS desktop', () => {
+  assert.equal(
+    resolveRunnerMaxConcurrentDestinationsFlag(MACOS_DEVICE),
+    '-maximum-concurrent-test-device-destinations',
+  );
+});
+
+test('resolveRunnerSigningBuildSettings returns empty args without env overrides', () => {
+  assert.deepEqual(resolveRunnerSigningBuildSettings({}), []);
+});
+
+test('resolveRunnerSigningBuildSettings disables signing for macOS desktop builds', () => {
+  assert.deepEqual(
+    resolveRunnerSigningBuildSettings({}, true, { platform: 'apple', appleOs: 'macos' }),
+    [
+      'CODE_SIGNING_ALLOWED=NO',
+      'CODE_SIGNING_REQUIRED=NO',
+      'CODE_SIGN_IDENTITY=',
+      'DEVELOPMENT_TEAM=',
+    ],
+  );
+});
+
+test('resolveRunnerSigningBuildSettings enables automatic signing for device builds without forcing identity', () => {
+  assert.deepEqual(resolveRunnerSigningBuildSettings({}, true), ['CODE_SIGN_STYLE=Automatic']);
+});
+
+test('resolveRunnerSigningBuildSettings ignores device signing overrides for simulator builds', () => {
+  assert.deepEqual(
+    resolveRunnerSigningBuildSettings(
+      {
+        AGENT_DEVICE_IOS_TEAM_ID: 'ABCDE12345',
+        AGENT_DEVICE_IOS_SIGNING_IDENTITY: 'Apple Development',
+        AGENT_DEVICE_IOS_PROVISIONING_PROFILE: 'My Profile',
+      },
+      false,
+    ),
+    [],
+  );
+});
+
+test('resolveRunnerSigningBuildSettings applies optional overrides when provided', () => {
+  const settings = resolveRunnerSigningBuildSettings(
+    {
+      AGENT_DEVICE_IOS_TEAM_ID: 'ABCDE12345',
+      AGENT_DEVICE_IOS_SIGNING_IDENTITY: 'Apple Development',
+      AGENT_DEVICE_IOS_PROVISIONING_PROFILE: 'My Profile',
+    },
+    true,
+  );
+  assert.deepEqual(settings, [
+    'CODE_SIGN_STYLE=Manual',
+    'DEVELOPMENT_TEAM=ABCDE12345',
+    'CODE_SIGN_IDENTITY=Apple Development',
+    'PROVISIONING_PROFILE_SPECIFIER=My Profile',
+  ]);
+});
+
+test('resolveRunnerSigningBuildSettings switches to manual signing when a profile is set without team or identity', () => {
+  const settings = resolveRunnerSigningBuildSettings(
+    { AGENT_DEVICE_IOS_PROVISIONING_PROFILE: 'My Profile' },
+    true,
+  );
+  assert.deepEqual(settings, [
+    'CODE_SIGN_STYLE=Manual',
+    'PROVISIONING_PROFILE_SPECIFIER=My Profile',
+  ]);
+});
+
+test('resolveRunnerPerformanceBuildSettings disables indexing and code coverage', () => {
+  assert.deepEqual(resolveRunnerPerformanceBuildSettings(), [
+    'COMPILER_INDEX_STORE_ENABLE=NO',
+    'ENABLE_CODE_COVERAGE=NO',
+    'ONLY_ACTIVE_ARCH=YES',
+    'ENABLE_PREVIEWS=NO',
+    'ENABLE_DEBUG_DYLIB=NO',
+  ]);
+});
+
+test('resolveRunnerSandboxBuildArgs disables nested Xcode and Swift sandboxing', () => {
+  assert.deepEqual(resolveRunnerSandboxBuildArgs(), [
+    '-IDEPackageSupportDisableManifestSandbox=1',
+    '-IDEPackageSupportDisablePluginExecutionSandbox=1',
+    'ENABLE_USER_SCRIPT_SANDBOXING=NO',
+    'OTHER_SWIFT_FLAGS=$(inherited) -disable-sandbox',
+  ]);
+});
+
+test('resolveRunnerSandboxBuildArgs includes Swift runner unit tests only when requested', () => {
+  const previous = process.env.AGENT_DEVICE_XCUITEST_INCLUDE_UNIT_TESTS;
+  try {
+    process.env.AGENT_DEVICE_XCUITEST_INCLUDE_UNIT_TESTS = '1';
+    assert.deepEqual(resolveRunnerSandboxBuildArgs(), [
+      '-IDEPackageSupportDisableManifestSandbox=1',
+      '-IDEPackageSupportDisablePluginExecutionSandbox=1',
+      'ENABLE_USER_SCRIPT_SANDBOXING=NO',
+      'OTHER_SWIFT_FLAGS=$(inherited) -disable-sandbox -D AGENT_DEVICE_RUNNER_UNIT_TESTS',
+    ]);
+  } finally {
+    if (previous === undefined) {
+      delete process.env.AGENT_DEVICE_XCUITEST_INCLUDE_UNIT_TESTS;
+    } else {
+      process.env.AGENT_DEVICE_XCUITEST_INCLUDE_UNIT_TESTS = previous;
+    }
+  }
+});
+
+test('resolveRunnerBundleBuildSettings returns default bundle identifiers', () => {
+  assert.deepEqual(resolveRunnerBundleBuildSettings({}), [
+    'AGENT_DEVICE_IOS_RUNNER_APP_BUNDLE_ID=com.callstack.agentdevice.runner',
+    'AGENT_DEVICE_IOS_RUNNER_TEST_BUNDLE_ID=com.callstack.agentdevice.runner.uitests',
+  ]);
+});
+
+test('resolveRunnerBundleBuildSettings uses AGENT_DEVICE_IOS_BUNDLE_ID when provided', () => {
+  assert.deepEqual(
+    resolveRunnerBundleBuildSettings({
+      AGENT_DEVICE_IOS_BUNDLE_ID: 'com.example.agent-device.runner',
+    }),
+    [
+      'AGENT_DEVICE_IOS_RUNNER_APP_BUNDLE_ID=com.example.agent-device.runner',
+      'AGENT_DEVICE_IOS_RUNNER_TEST_BUNDLE_ID=com.example.agent-device.runner.uitests',
+    ],
+  );
+});
